converted to double precision - succesful simulation
git-svn-id: https://svn.oss.deltares.nl/repos/imod/trunk@1867 8a8853d3-af7d-47ec-b359-4bd2096ba547

+++ MODFLOW/MF2005/gwf2bcf7.f
@@ -43,7 +43,7 @@ c   If not, see <http://water.usgs.gov/software/help/notice/>.
        REAL, SAVE,     POINTER,   DIMENSION(:,:,:) ::SC2
        REAL, SAVE,     POINTER,   DIMENSION(:,:,:) ::WETDRY
        REAL, SAVE,     POINTER,   DIMENSION(:,:,:) ::CVWD
-       REAL, SAVE,     POINTER,   DIMENSION(:)     ::TRPY
+       REAL(kind=8), SAVE,     POINTER,   DIMENSION(:)     ::TRPY
        real, save,     pointer,   dimension(:,:)   ::seepage            ! DLT
 
       TYPE GWFBCFTYPE
@@ -58,7 +58,7 @@ c   If not, see <http://water.usgs.gov/software/help/notice/>.
        REAL,     POINTER,   DIMENSION(:,:,:) ::SC2
        REAL,     POINTER,   DIMENSION(:,:,:) ::WETDRY
        REAL,     POINTER,   DIMENSION(:,:,:) ::CVWD
-       REAL,     POINTER,   DIMENSION(:)     ::TRPY
+       REAL(kind=8),     POINTER,   DIMENSION(:)     ::TRPY
        real,     pointer,   dimension(:,:)   ::seepage                  ! DLT
       END TYPE
       TYPE(GWFBCFTYPE), SAVE  ::GWFBCFDAT(10)
